cosmetic tweaks
