+++ client/src/pages/Register/index.js
@@ -2,6 +2,8 @@ import React, { Component } from "react";
 import { Col, Row, Container } from "../../components/Grid";
 import { Input } from "../../components/Form";
 import Jumbotron from "../../components/Jumbotron";
+import Navbar from "../../components/Navbar";
+
 import "./style.css";
 import axios from "axios";
 
@@ -63,6 +65,7 @@ class Register extends Component {
     return (
       <Container fluid>
         <Row>
+          <Navbar />
           <Col size="md-12">
             <Jumbotron>
               <h1>Register</h1>
