+++ server.js
@@ -4,7 +4,6 @@ var session = require("express-session");
 // Requiring passport as we've configured it
 const logger = require("morgan");
 require("dotenv").config();
-const mysql = require("mysql");
 const Cors = require("cors");
 const passport = require("passport");
 require("./config/passport");
@@ -12,7 +11,6 @@ require("./config/passport");
 // Setting up port and requiring models for syncing
 var PORT = process.env.PORT || 3001;
 var db = require("./models");
-const authRoutes = require("./routes/auth-routes");
 // Creating express app and configuring middleware needed for authentication
 var app = express();
 app.use(Cors());
@@ -38,25 +36,9 @@ app.use(
 );
 app.use(passport.initialize());
 app.use(passport.session());
-app.use(routes);
+
 app.use(logger("combined"));
 
-//connection = require("./config/connection");
-
-// connection.connect(function(err) {
-//   if (err) {
-//     console.error("error connecting: " + err.stack);
-//     return;
-//   }
-//   app.listen(PORT, function() {
-//     console.log(
-//       "==> :earth_americas:  Listening on port %s. Visit http://localhost:%s/ in your browser.",
-//       PORT,
-//       PORT
-//     );
-//   });
-//   console.log("connected as id " + connection.threadId);
-// });
 //Syncing our database and logging a message to the user upon success
 db.sequelize.sync({ force: false }).then(function() {
   app.listen(PORT, function() {
